* lib: extended QFitFunction with functions to estimate a specific parameter from an XY-location in a plot (also supported by library fit functions in a DLL/SOLIB...) * curve_fit/fcs_fit/imfcs_fit: added context-menu that allows to estimate fit parameters from the mouse-position * fcs_fitfunctions: added code for single-parameter-estimation (from XY) and initial-parameter guessing * diverse code-refactorings in allmost all fit functions (derived from general base classes that implement generic rules for parameter estimation in many reoccuring cases)

+++ plugins/fcs/qfrdrfcsdata.cpp
@@ -981,7 +981,7 @@ bool QFRDRFCSData::loadCorrelationCurvesFromCSV(QStringList filenames) {
                         }
                         runincrement=(columns-1)/2;
                     } else if (mode==2) { // tau, corr, corr, ... -> select only one column
-                        if (column<data[ii].tab->get_column_count()) correlation[(run0)*correlationN+l]=data[ii].tab->get(column, l);
+                        if (column<(long long)data[ii].tab->get_column_count()) correlation[(run0)*correlationN+l]=data[ii].tab->get(column, l);
                         runincrement=1;
                     }
                 }
